tourism2/requirements.txt
streamlit
mysql-connector-python
google-generativeai
pandas
plotly